backend_old/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic
uvicorn
python-dotenv
pytest